Refactor: rename package
