+++ src/main/java/subway/domain/EndStations.java
@@ -1,4 +1,4 @@
-package subway.entity;
+package subway.domain;
 
 import java.util.Set;
 import javax.persistence.Entity;
