+++ src/main/java/subway/repository/StationRepository.java
@@ -1,7 +1,7 @@
 package subway.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.entity.Station;
+import subway.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
