+++ src/main/java/subway/service/LineService.java
@@ -12,9 +12,9 @@ import subway.controller.dto.LineRequest;
 import subway.controller.dto.LineRequest.UpdateRequest;
 import subway.controller.dto.LineResponse;
 import subway.controller.dto.LinesResponse;
-import subway.entity.EndStations;
-import subway.entity.Line;
-import subway.entity.Station;
+import subway.domain.EndStations;
+import subway.domain.Line;
+import subway.domain.Station;
 import subway.exception.LineNotFoundException;
 import subway.exception.StationNotFoundException;
 import subway.repository.LineRepository;
